requirements.txt
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
aiogram==3.13.1
asyncpg==0.29.0
sqlalchemy==2.0.31
aiohttp==3.9.5
aiocache==0.12.2
uvloop==0.19.0
